+++ src/dataPublisherBase.h
@@ -150,6 +150,10 @@ class dataPublisher {
      * @param baseLogger The logger supplying the data to be published
      */
     void begin(Logger& baseLogger);
+    /**
+     * @brief Begin the publisher by doing the minimum that is needed after boot
+     */
+    virtual void begin();
 
 
     /**
@@ -162,6 +166,14 @@ class dataPublisher {
 
 
     /**
+     * @brief Checks if the publisher needs an Internet connection for the next
+     * publishData call (as opposed to just buffering data internally).
+     *
+     * @return True if an internet connection is needed for the next publish.
+     */
+    virtual bool connectionNeeded(void);
+
+    /**
      * @brief Open a socket to the correct receiver and sends out the formatted
      * data.
      *
@@ -171,10 +183,11 @@ class dataPublisher {
      * @param outClient An Arduino client instance to use to print data to.
      * Allows the use of any type of client and multiple clients tied to a
      * single TinyGSM modem instance
+     * @param forceFlush Ask the publisher to flush buffered data immediately.
      * @return The result of publishing data.  May be an http
      * response code or a result code from PubSubClient.
      */
-    virtual int16_t publishData(Client* outClient) = 0;
+    virtual int16_t publishData(Client* outClient, bool forceFlush = false) = 0;
     /**
      * @brief Open a socket to the correct receiver and send out the formatted
      * data.
@@ -183,10 +196,12 @@ class dataPublisher {
      * either a client having been linked to the publisher or a logger modem
      * having been linked to the logger linked to the publisher.
      *
+     * @param forceFlush Ask the publisher to flush buffered data immediately.
+     *
      * @return The result of publishing data.  May be an http
      * response code or a result code from PubSubClient.
      */
-    virtual int16_t publishData();
+    virtual int16_t publishData(bool forceFlush = false);
 
     /**
      * @brief Retained for backwards compatibility; use publishData(Client*