+++ src/publishers/DreamHostPublisher.h
@@ -134,10 +134,11 @@ class DreamHostPublisher : public dataPublisher {
      * @param outClient An Arduino client instance to use to print data to.
      * Allows the use of any type of client and multiple clients tied to a
      * single TinyGSM modem instance
+     * @param forceFlush Ask the publisher to flush buffered data immediately.
      *
      * @return The http status code of the response.
      */
-    int16_t publishData(Client* outClient) override;
+    int16_t publishData(Client* outClient, bool forceFlush = false) override;
 
  protected:
     // portions of the GET request